requirements.txt
pytest
PyYAML
markdown
python-frontmatter
black
flake8
rich
pytest-cov
